requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-taggit==1.3.0
django-taggit-templatetags2==1.6.1
gunicorn==20.0.4
importlib-metadata==2.0.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
soupsieve==2.0.1
sqlparse==0.4.1
timeago==1.0.15
whitenoise==5.2.0
zipp==3.4.0
